+++ DTCD-StyleSystem/src/base-components/BaseTooltip/BaseTooltip.js
@@ -4,8 +4,11 @@ import styles from './BaseTooltip.scss';
 export default class BaseTooltip extends HTMLElement {
 
   #placements = ['top', 'bottom', 'left', 'right'];
-  #tooltip;
+  #placement;
   #content;
+
+  #container;
+  #tooltip;
 
   static get observedAttributes() {
     return ['content', 'placement'];
@@ -24,9 +27,10 @@ export default class BaseTooltip extends HTMLElement {
     this.shadowRoot.appendChild(style);
     style.appendChild(document.createTextNode(styles));
 
-    this.#tooltip = this.shadowRoot.querySelector('#tooltip');
-    this.#tooltip.addEventListener('animationstart', this.#animationStartHandler);
-    this.#tooltip.addEventListener('animationend', this.#animationEndHandler);
+    this.#container = this.shadowRoot.querySelector('.Container');
+
+    this.#tooltip = document.createElement('div');
+    this.#tooltip.classList.add('Tooltip');
   }
 
   get content() {
@@ -49,32 +53,42 @@ export default class BaseTooltip extends HTMLElement {
     }
   }
 
-  #setPlacement(value = 'top') {
-    const placement = this.#placements.includes(value) ? value : 'top';
-    this.#tooltip.classList = `Tooltip ${placement}`;
-  }
-
-  #animationStartHandler(e) {
-    if (e.animationName === 'fade-in') {
-      e.target.classList.add('did-fade-in');
-    }
-  }
-
-  #animationEndHandler(e) {
-    if (e.animationName === 'fade-out') {
-      e.target.classList.remove('did-fade-in');
-    }
-  }
-
   connectedCallback() {
+    this.addEventListener('mouseenter', this.#handlerHoverStart);
+    this.addEventListener('mouseleave', this.#handlerHoverEnd);
+
+    const options = {
+      root: document.body,
+      rootMargin: "0px",
+      threshold: 1.0,
+    };
+
+    const callback = (entries, observer) => {
+      entries.forEach((entry) => {
+        // console.log(entry.boundingClientRect);
+        // Each entry describes an intersection change for one observed
+        // target element:
+        //   entry.boundingClientRect
+        //   entry.intersectionRatio
+        //   entry.intersectionRect
+        //   entry.isIntersecting
+        //   entry.rootBounds
+        //   entry.target
+        //   entry.time
+      });
+    };
+    
+    const observer = new IntersectionObserver(callback, options);
+          observer.observe(this);
+
     if (!this.hasAttribute('placement')) {
       this.#setPlacement('top');
     }
   }
 
   disconnectedCallback() {
-    this.#tooltip.removeEventListener('animationstart', this.#animationStartHandler);
-    this.#tooltip.removeEventListener('animationend', this.#animationEndHandler);
+    this.removeEventListener('mouseenter', this.#handlerHoverStart);
+    this.removeEventListener('mouseleave', this.#handlerHoverEnd);
   }
 
   attributeChangedCallback(attrName, oldValue, newValue) {
@@ -87,4 +101,86 @@ export default class BaseTooltip extends HTMLElement {
     }
   }
 
+  #setPlacement(value = 'top') {
+    this.#placement = this.#placements.includes(value) ? value : 'top';
+  }
+
+  #setPosition() {
+    const {
+      height,
+      width,
+      left,
+      top,
+    } = this.getBoundingClientRect();
+
+    const render = (hintPos) => {
+      const {
+        style,
+        classList,
+      } = this.#tooltip;
+
+      classList.remove('top', 'bottom', 'left', 'right');
+      classList.add(hintPos);
+  
+      switch (hintPos) {
+        case 'bottom':
+          style.top = (top + height) + 'px';
+          style.left = (left + width / 2) + 'px';
+          break;
+  
+        case 'top':
+          style.top = top + 'px';
+          style.left = (left + width / 2) + 'px';
+          break;
+  
+        case 'left':
+          style.top = (top + height / 2) + 'px';
+          style.left = left + 'px';
+          break;
+  
+        case 'right':
+          style.top = (top + height / 2) + 'px';
+          style.left = (left + width) + 'px';
+          break;
+      
+        default:
+          break;
+      }
+    }
+
+    if (!this.placement) {
+      this.placement = this.#placements[0];
+    }
+    render(this.placement);
+
+    const {
+      offsetHeight: windowHeight,
+      offsetWidth: windowWidth,
+    } = document.body;
+    const {
+      top: hintTop,
+      bottom: hintBottom,
+      left: hintLeft,
+      right: hintRight,
+    } = this.#tooltip.getBoundingClientRect();
+
+    let posAwayFromScreen = this.#placement;
+    if (hintTop < 0) posAwayFromScreen = 'bottom';
+    if (hintBottom > windowHeight) posAwayFromScreen = 'top';
+    if (hintLeft < 0) posAwayFromScreen = 'right';
+    if (hintRight > windowWidth) posAwayFromScreen = 'left';
+
+    if (this.placement !== posAwayFromScreen) {
+      render(posAwayFromScreen);
+    }
+  }
+
+  #handlerHoverStart() {
+    this.#container.append(this.#tooltip);
+    this.#setPosition();
+  }
+
+  #handlerHoverEnd() {
+    this.#tooltip.remove();
+  }
 }