+++ DTCD-StyleSystem/.storybook/main.js
@@ -1,56 +1,53 @@
 const path = require('path');
-
 module.exports = {
-  core: {
-    builder: 'webpack5',
-  },
   'stories': [
     '../src/**/*.stories.mdx',
-    '../src/**/*.stories.@(js|jsx|ts|tsx)'
+    '../src/**/*.stories.@(js|jsx|ts|tsx)',
   ],
   'addons': [
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/manager-webpack5',
     'storybook-dark-mode',
+    '@storybook/addon-mdx-gfm',
   ],
-  framework: '@storybook/web-components',
-
-  webpackFinal: async (config, { configType }) => {
+  framework: {
+    name: '@storybook/web-components-webpack5',
+    options: {}
+  },
+  webpackFinal: async (config, {
+    configType
+  }) => {
     // `configType` has a value of 'DEVELOPMENT' or 'PRODUCTION'
     // You can change the configuration based on that.
     // 'PRODUCTION' is used when building the static version of storybook.
 
-    config.module.rules.forEach((rule) => {
+    config.module.rules.forEach(rule => {
       if (rule.test && '.css'.match(rule.test)) {
         rule.exclude = path.resolve(__dirname, '../src/base-components');
       }
     });
-
     config.module.rules.push({
       test: /\.html$/i,
       include: path.resolve(__dirname, '../src/base-components'),
-      use: ['html-loader'],
+      use: ['html-loader']
     });
-
     config.module.rules.push({
       test: /\.((c|sa|sc)ss)$/i,
       include: path.resolve(__dirname, '../src/base-components'),
-      use: [
-        {
-          loader: 'css-loader',
-          options: {
-            exportType: 'string',
-            sourceMap: false,
-          }
-        },
-        {
-          loader: 'postcss-loader',
-        },
-        'sass-loader'
-      ],
+      use: [{
+        loader: 'css-loader',
+        options: {
+          exportType: 'string',
+          sourceMap: false
+        }
+      }, {
+        loader: 'postcss-loader'
+      }, 'sass-loader']
     });
-
     return config;
   },
-}
+  docs: {
+    autodocs: true
+  }
+};