+++ DTCD-StyleSystem/src/stories/BaseTabs.stories.js
@@ -4,6 +4,11 @@ import BaseTabsDoc from './docs/BaseTabsDoc.mdx';
 
 export default {
   title: 'Example/BaseComponents/BaseTabs',
+  argTypes: {
+    activeTab: {
+      type: 'number',
+    },
+  },
   parameters: {
     docs: {
       page: BaseTabsDoc,
@@ -18,9 +23,11 @@ window.customElements.define(NAME_COMPONENT, BaseTabs);
 const Template = (args) => {
   const {
     tabSlot = [],
+    activeTab,
   } = args;
 
   const tabs = document.createElement(NAME_COMPONENT);
+  tabs.activeTab = activeTab;
 
   tabSlot.forEach((slot) => {
     tabs.innerHTML += slot;
@@ -36,4 +43,5 @@ DefaultTabs.args = {
     `<span slot="tab">Tab 2 content</span>`,
     `<p slot="tab" tab-name="Test">Tab 3 content</p>`,
   ],
+  activeTab: 1,
 };