+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - toggle event in BaseExpander
+- attribute `data-active-tab` in BaseTabs
 
 ## [0.13.0]