+++ CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - toggle event in BaseExpander
 - attribute `data-active-tab` in BaseTabs
 - closing the BaseDataTimePicker on the cancel button
+- variable in storybook global styles
 
 ### Changed
 
 - adding of label to BaseSelect
+- tooltip title color
 
 ### Fixed