+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - adding of label to BaseSelect
 - tooltip title color
+- tooltip rendering so that it doesn't get cut off
 
 ### Fixed